In most cases, modifying player amount works
When decrementing the player amount from 2 to 1, the wrong elements are deleted. These elements were not added by my function. Because the way I remove players is by deleting the last 3 elements in the grid, I think it's deleting the wrong ones because the order that Qt put them in the grid is different.